Add a symbolic name for the comparison in the goal & 3 types of sensors
When we defined a condition in a goal, instead of having a 'Temperature' hard coded, we can now choose it's own name in order to be easier for the one who will create a badge to know what it needs as a sensor. We can also choose between 3 types of sensors values (boolean, string, number) to propose a different form when we create the goal.